WIP: Particles!

+++ simple-invaders/src/sprites.rs
@@ -2,8 +2,9 @@ use std::cmp::min;
 use std::rc::Rc;
 use std::time::Duration;
 
+use crate::geo::Point;
 use crate::loader::Assets;
-use crate::{Point, SCREEN_HEIGHT, SCREEN_WIDTH};
+use crate::{SCREEN_HEIGHT, SCREEN_WIDTH};
 use line_drawing::Bresenham;
 
 // This is the type stored in the `Assets` hash map
@@ -187,6 +188,10 @@ impl Animation for SpriteRef {
 }
 
 /// Blit a drawable to the pixel buffer.
+///
+/// # Panics
+///
+/// Asserts that the drawable is fully contained within the screen.
 pub(crate) fn blit<S>(screen: &mut [u8], dest: &Point, sprite: &S)
 where
     S: Drawable,
@@ -215,8 +220,8 @@ where
 
 /// Draw a line to the pixel buffer using Bresenham's algorithm.
 pub(crate) fn line(screen: &mut [u8], p1: &Point, p2: &Point, color: [u8; 4]) {
-    let p1 = (p1.x as i64, p1.y as i64);
-    let p2 = (p2.x as i64, p2.y as i64);
+    let p1 = (p1.x as i32, p1.y as i32);
+    let p2 = (p2.x as i32, p2.y as i32);
 
     for (x, y) in Bresenham::new(p1, p2) {
         let x = min(x as usize, SCREEN_WIDTH - 1);
